refactor(sdk): move Ed25519 crypto primitives into SDK package (T2 PR 2 precursor)
The SDK is MIT-licensed and connector authors need to sign their own
manifests. Keeping the canonical-JSON serializer + Ed25519 verify/sign
primitives in the AGPL-licensed gateway package would either (a) force
connectors to take an AGPL dep, or (b) silently violate the
sdk-no-import-core dep-cruiser rule when the SDK re-exported gateway
symbols.

Solution: move both files to packages/sdk/src/crypto/. Gateway code
keeps importing through thin re-export shims at the old paths so
existing relative imports keep working with no further churn.
